requirements.txt
Flask
pillow
flask-cors